Fix edge case for class display names
Simply using the class display name in test logging does not work,
because JUnit platform returns only the simple class name from
TestIdentifier.getDisplayName(). This results in the fully qualified
package name being cut off for top level classes.

This change adds some logic that tried to find out whether the display
name has been explicitly se tor not. If so then the class display name
will we use. If not the fully qualified class name will be used for top
level classes and the simple class name will be used for inner classes.

There is still one corner case that is not supported with this solution:
A user could set the simple class name as display name in order to remove
the package name from the output. Since there is no way to tell whether
this was explicitly configured by the user or not by looking at the
TestIdentifier, there does not seem to be a way to support this.